requirements.txt
streamlit>=1.26.0
pandas>=2.0.0
mysql-connector-python>=8.0.33
sqlalchemy>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
matplotlib>=3.0.0
scipy>=1.10.0
